+++ src/lib/lisp/gc.c
@@ -13,6 +13,7 @@
 
 #ifdef TARGET_UNIX
 #include <stdio.h>
+#include <strings.h>
 #endif
 
 extern lispptr lisp_fnin;
@@ -294,6 +295,10 @@ check_xlat:
 #else // #ifdef FRAGMENTED_HEAP
     // Save free pointer.
     heap_free = d;
+#endif
+
+#ifndef NDEBUG
+    bzero (d, heap_end - d);
 #endif
 
     // End symbol list.